Refactor: Improve OpenCuaHandler logic for retry count and enhance message conversion validation
