+++ server/computer_use/handlers/registry.py
@@ -8,9 +8,9 @@ the appropriate handler based on the API provider.
 from typing import Dict, Optional, Type
 
 from server.computer_use.config import APIProvider
-from server.computer_use.handlers.anthropic import AnthropicHandler
+from server.computer_use.handlers.anthropic.handler import AnthropicHandler
 from server.computer_use.handlers.base import BaseProviderHandler, ProviderHandler
-from server.computer_use.handlers.openai import OpenAIHandler
+from server.computer_use.handlers.openai.handler import OpenAIHandler
 from server.computer_use.handlers.opencua.handler import OpenCuaHandler
 
 # Registry mapping providers to handler classes
